feat(gpu): add W5f table transfer and blend filters
Reuse the shared conversion and blend formula authorities, preserve typed U32 Table payloads through the V4 packing permit, and add bounded public Rect/Path witnesses for all 29 blend modes.

Verification: 64 public assertions pass; native test command remains failed (executor 43 exit 133). Standalone five-module compile succeeds. R15 readonly UBytes and R16 typed transport remain scoped; no root refactor documents included.

+++ gpu-renderer/src/main/kotlin/org/graphiks/kanvas/gpu/renderer/materials/W5aMaterialSourceStage.kt
@@ -54,7 +54,7 @@ internal class W5aMaterialSourceStage private constructor(
             if (requirements.uniformByteCountI64 % 16L != 0L || wordsI64 !in 1L..Int.MAX_VALUE.toLong()) return null
             val code = W5fColorOperationEmitterV1.emit(proof.copyOperationGraph(),"vec4<f32>(0.0)",0L)
             return W5aMaterialSourceStage(requirements,"""
-                struct W5fMaterialBlock { words: array<vec4<f32>, ${wordsI64}>, }
+                struct W5fMaterialBlock { words: array<vec4<u32>, ${wordsI64}>, }
                 @group(1) @binding(0) var<uniform> w5fMaterial: W5fMaterialBlock;
                 fn kanvas_material_source(localPosition: vec2<f32>) -> vec4<f32> {
                     $code
